refactor: 외부 모달 클릭으로 인해, 스크린 리더 상태를 변화시키는 로직 커스텀 훅으로 분리
